@@ -32,6 +32,7 @@ class ProgramsController < ApplicationController
   end
 
   def create
+    flash = {}
     @program = Program.new(params[:program])
 
     if @program.save
@@ -68,6 +69,7 @@ class ProgramsController < ApplicationController
   end
 
   def update 
+    flash = {}
     @program = Program.find(params[:id])
     if @program.update_attributes(params[:program])
       flash[:notice] = "Programa actualizada."
@@ -111,6 +113,7 @@ class ProgramsController < ApplicationController
   end
 
   def create_course
+    flash = {}
     @program = Program.find(params[:program_id])
     if @program.update_attributes(params[:program])
       flash[:notice] = "Nuevo curso creado."
@@ -132,6 +135,7 @@ class ProgramsController < ApplicationController
   end
 
   def create_term
+    flash = {}
     @program = Program.find(params[:program_id])
     if @program.update_attributes(params[:program])
       flash[:notice] = "Nuevo curso creado."
@@ -175,6 +179,7 @@ class ProgramsController < ApplicationController
   end
 
   def create_enrollment
+    flash = {}
     @term = Term.find(params[:term_id])
     if @term.update_attributes(params[:term])
       flash[:notice] = "Estudiante inscrito satisfactoriamente"
@@ -192,6 +197,7 @@ class ProgramsController < ApplicationController
   end
 
   def update_enrollment
+    flash = {}
     @ts = TermStudent.find(params[:ts][:id])
     if @ts.update_attributes(params[:ts])
       flash[:notice] = "Inscripción actualizada"
@@ -270,6 +276,7 @@ class ProgramsController < ApplicationController
   end
 
   def create_schedule
+    flash = {}
     @tc = TermCourse.find(params[:term_course_id])
     if @tc.update_attributes(params[:term_course])
       flash[:notice] = "Sesión creada."
@@ -288,6 +295,7 @@ class ProgramsController < ApplicationController
   end
 
   def update_schedule
+    flash = {}
     @cs = TermCourseSchedule.find(params[:cs][:id])
     if @cs.update_attributes(params[:cs])
       if (@cs.status == TermCourseSchedule::INACTIVE) 
@@ -373,6 +381,7 @@ class ProgramsController < ApplicationController
   end
 
   def create_course_student
+    flash = {}
     @tc = TermCourse.find(params[:term_course_id])
     if @tc.update_attributes(params[:term_course])
       flash[:notice] = "Estudiante agregado."
@@ -390,6 +399,7 @@ class ProgramsController < ApplicationController
   end
 
   def update_course_student
+    flash = {}
     @cs = TermCourseStudent.find(params[:cs][:id])
     if @cs.update_attributes(params[:cs])
       flash[:notice] = "Estudiante actualizado."
@@ -423,6 +433,7 @@ class ProgramsController < ApplicationController
   end
 
   def inactive_course_student
+    flash = {}
     @cs = TermCourseStudent.find(params[:term_course_student_id])
     params[:cs] = {:status => TermCourseStudent::INACTIVE}
     if @cs.update_attributes(params[:cs])
@@ -495,6 +506,7 @@ class ProgramsController < ApplicationController
   end
 
   def create_group
+    flash = {}
     @tc = TermCourse.new
     if @tc.update_attributes(params[:term_course])
       flash[:notice] = "Grupo creado."
@@ -510,6 +522,7 @@ class ProgramsController < ApplicationController
   end
 
   def update_group
+    flash = {}
     @tc = TermCourse.find(params[:term_course][:id])
     if @tc.update_attributes(params[:term_course])
       flash[:notice] = "El titular del grupo ha sido actualizado."
@@ -544,8 +557,9 @@ class ProgramsController < ApplicationController
   end
   
   def upload_file
+    flash = {}
     params[:documentation_file]['file'].each do |f|
-      @documentation_file = DocumentationFile.new(f)
+      @documentation_file = DocumentationFile.new
       @documentation_file.program_id = params[:documentation_file]['program_id']
       @documentation_file.file = f
       @documentation_file.description = f.original_filename
