@@ -3,7 +3,7 @@ class InternshipFileUploader < CarrierWave::Uploader::Base
   storage :file
 
   def store_dir
-    "#{RAILS_ROOT}/private/files/internships/#{model.internship_id}"
+    "#{Rails.root}/private/files/internships/#{model.internship_id}"
   end
 
 end
