@@ -13,7 +13,7 @@ class StudentFileUploader < CarrierWave::Uploader::Base
   # Override the directory where uploaded files will be stored.
   # This is a sensible default for uploaders that are meant to be mounted:
   def store_dir
-    "#{RAILS_ROOT}/private/files/students/#{model.student_id}"
+    "#{Rails.root}/private/files/students/#{model.student_id}"
   end
 
   # Provide a default URL as a default if there hasn't been a file uploaded:
