@@ -1,6 +1,3 @@
-require 'openid/store/filesystem'
-
 Rails.application.config.middleware.use OmniAuth::Builder do
-  provider :google_apps, OpenID::Store::Filesystem.new('/tmp'), :domain => GAPPS_DOMAIN, :name => 'admin'
+  provider :google_oauth2, ENV['GOOGLE_KEY'], ENV['GOOGLE_SECRET']
 end
-
