@@ -1,3 +1,4 @@
+# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -306,8 +307,8 @@ ActiveRecord::Schema.define(:version => 20120607225526) do
     t.integer  "staff_id"
     t.string   "description"
     t.string   "file"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.datetime "created_at",  :null => false
+    t.datetime "updated_at",  :null => false
   end
 
   add_index "staff_files", ["staff_id"], :name => "index_staff_files_on_staff_id"
